Remove created & modfied times from State
Using created and modified times is not reliable on certain Linux systems.
Previously, the `test_changed_content_rewritten` test in `test_state_has_changed`
would fail on Linux because the they would not change when the test rewrote the
same content into the file.

This change also means that the file is considered to be "unchanged" if the
contents of the file was changed back.

Signed-off-by: Kah Goh <villastar@yahoo.com.au>